% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nina}
\alias{nina}
\title{nina}
\format{
A data frame with 74 rows and 4 variables:
\describe{
\item{\code{ID}}{double Participant ID}
\item{\code{reputational_harm}}{double Perceived harm to the reputation of a fictional adversary (Nina)}
\item{\code{selfclose}}{double Sum of participant's self-rated closeness with four closest friends in the sorority}
\item{\code{friendsclose}}{double Sum of participant's perception of the closeness of her four best friends in the sorority}
}
}
\usage{
nina
}
\description{
Sorority data on real-life social network closeness and reputation of fictional adversary.
}
\details{
Data collected from a sorority in southern California in the early 2000's.
}
\keyword{datasets}
